Copter: Fix a couple of px4 build warnings

+++ ArduCopter/Log.pde
@@ -182,7 +182,7 @@ static void Log_Write_AutoTune(uint8_t axis, uint8_t tune_step, float rate_min, 
         new_gain_rd  : new_gain_rd,
         new_gain_sp  : new_gain_sp
     };
-    DataFlash.WriteBlock(&pkt, sizeof(pkt));
+    DataFlash.WriteBlock(&pkt, (uint16_t) sizeof(pkt));
 }
 
 struct PACKED log_AutoTuneDetails {
@@ -199,7 +199,7 @@ static void Log_Write_AutoTuneDetails(int16_t angle_cd, float rate_cds)
         angle_cd    : angle_cd,
         rate_cds    : rate_cds
     };
-    DataFlash.WriteBlock(&pkt, sizeof(pkt));
+    DataFlash.WriteBlock(&pkt, (uint16_t) sizeof(pkt));
 }
 #endif
 
@@ -227,7 +227,7 @@ static void Log_Write_Current()
         board_voltage       : board_voltage(),
         current_total       : battery.current_total_mah()
     };
-    DataFlash.WriteBlock(&pkt, sizeof(pkt));
+    DataFlash.WriteBlock(&pkt, (uint16_t) sizeof(pkt));
 }
 
 struct PACKED log_Optflow {
@@ -255,7 +255,7 @@ static void Log_Write_Optflow()
         roll            : of_roll,
         pitch           : of_pitch
     };
-    DataFlash.WriteBlock(&pkt, sizeof(pkt));
+    DataFlash.WriteBlock(&pkt, (uint16_t) sizeof(pkt));
  #endif     // OPTFLOW == ENABLED
 }
 
@@ -295,7 +295,7 @@ static void Log_Write_Nav_Tuning()
         desired_accel_x : wp_nav.desired_accel.x,
         desired_accel_y : wp_nav.desired_accel.y
     };
-    DataFlash.WriteBlock(&pkt, sizeof(pkt));
+    DataFlash.WriteBlock(&pkt, (uint16_t) sizeof(pkt));
 }
 
 struct PACKED log_Control_Tuning {
@@ -330,7 +330,7 @@ static void Log_Write_Control_Tuning()
         desired_climb_rate  : desired_climb_rate,
         climb_rate          : climb_rate
     };
-    DataFlash.WriteBlock(&pkt, sizeof(pkt));
+    DataFlash.WriteBlock(&pkt, (uint16_t) sizeof(pkt));
 }
 
 struct PACKED log_Compass {
@@ -385,7 +385,7 @@ static void Log_Write_Compass()
             motor_offset_y  : (int16_t)mag2_motor_offsets.y,
             motor_offset_z  : (int16_t)mag2_motor_offsets.z
         };
-        DataFlash.WriteBlock(&pkt2, sizeof(pkt2));
+        DataFlash.WriteBlock(&pkt2, (uint16_t) sizeof(pkt2));
     }
 #endif
 }
@@ -418,7 +418,7 @@ static void Log_Write_Performance()
         ins_error_count  : ins.error_count(),
         inav_error_count : inertial_nav.error_count()
     };
-    DataFlash.WriteBlock(&pkt, sizeof(pkt));
+    DataFlash.WriteBlock(&pkt, (uint16_t) sizeof(pkt));
 }
 
 struct PACKED log_Cmd {
@@ -447,7 +447,7 @@ static void Log_Write_Cmd(uint8_t num, const struct Location *wp)
         waypoint_latitude   : wp->lat,
         waypoint_longitude  : wp->lng
     };
-    DataFlash.WriteBlock(&pkt, sizeof(pkt));
+    DataFlash.WriteBlock(&pkt, (uint16_t) sizeof(pkt));
 }
 
 struct PACKED log_Attitude {
@@ -491,7 +491,7 @@ static void Log_Write_Mode(uint8_t mode)
         mode            : mode,
         throttle_cruise : g.throttle_cruise,
     };
-    DataFlash.WriteBlock(&pkt, sizeof(pkt));
+    DataFlash.WriteBlock(&pkt, (uint16_t) sizeof(pkt));
 }
 
 struct PACKED log_Startup {
@@ -504,7 +504,7 @@ static void Log_Write_Startup()
     struct log_Startup pkt = {
         LOG_PACKET_HEADER_INIT(LOG_STARTUP_MSG)
     };
-    DataFlash.WriteBlock(&pkt, sizeof(pkt));
+    DataFlash.WriteBlock(&pkt, (uint16_t) sizeof(pkt));
 }
 
 struct PACKED log_Event {
@@ -520,7 +520,7 @@ static void Log_Write_Event(uint8_t id)
             LOG_PACKET_HEADER_INIT(LOG_EVENT_MSG),
             id  : id
         };
-        DataFlash.WriteBlock(&pkt, sizeof(pkt));
+        DataFlash.WriteBlock(&pkt, (uint16_t) sizeof(pkt));
     }
 }
 
@@ -539,7 +539,7 @@ static void Log_Write_Data(uint8_t id, int16_t value)
             id          : id,
             data_value  : value
         };
-        DataFlash.WriteBlock(&pkt, sizeof(pkt));
+        DataFlash.WriteBlock(&pkt, (uint16_t) sizeof(pkt));
     }
 }
 
@@ -558,7 +558,7 @@ static void Log_Write_Data(uint8_t id, uint16_t value)
             id          : id,
             data_value  : value
         };
-        DataFlash.WriteBlock(&pkt, sizeof(pkt));
+        DataFlash.WriteBlock(&pkt, (uint16_t) sizeof(pkt));
     }
 }
 
@@ -577,7 +577,7 @@ static void Log_Write_Data(uint8_t id, int32_t value)
             id          : id,
             data_value  : value
         };
-        DataFlash.WriteBlock(&pkt, sizeof(pkt));
+        DataFlash.WriteBlock(&pkt, (uint16_t) sizeof(pkt));
     }
 }
 
@@ -596,7 +596,7 @@ static void Log_Write_Data(uint8_t id, uint32_t value)
             id          : id,
             data_value  : value
         };
-        DataFlash.WriteBlock(&pkt, sizeof(pkt));
+        DataFlash.WriteBlock(&pkt, (uint16_t) sizeof(pkt));
     }
 }
 
@@ -615,7 +615,7 @@ static void Log_Write_Data(uint8_t id, float value)
             id          : id,
             data_value  : value
         };
-        DataFlash.WriteBlock(&pkt, sizeof(pkt));
+        DataFlash.WriteBlock(&pkt, (uint16_t) sizeof(pkt));
     }
 }
